fixing docker issues not catching sigterm kills

+++ tests/exchange/test_exchange_adapter.py
@@ -1,7 +1,7 @@
 import os
 import pandas as pd
 import pytest
-from core.exchanges.exchange_adapter import ExchangeAdapter, create_base_folder
+from core.exchanges.exchange_adapter import ExchangeAdapter
 
 @pytest.fixture
 def adapter_validate_message_is_false():
@@ -15,9 +15,7 @@ def adapter_validate_message_is_false():
             return batch_list
 
 
-    path_to_folder = r"D:\python_projects\data_gathering_using_websockets_finance_crypto\data"
-
-    return TestAdapter(path_to_folder=path_to_folder, exchange_name="Test", url="ws:test", msg={}, ticker="TEST_TEST")
+    return TestAdapter(exchange_name="Test", url="ws:test", msg={}, ticker="TEST_TEST")
 
 
 @pytest.fixture
@@ -31,10 +29,7 @@ def adapter():
         def normalise_data(self, batch_list:list) -> list:
             return batch_list
 
-
-    path_to_folder = r"D:\python_projects\data_gathering_using_websockets_finance_crypto\data"
-
-    return TestAdapter(path_to_folder=path_to_folder, exchange_name="Test", url="ws:test", msg={}, ticker="TEST_TEST")
+    return TestAdapter(exchange_name="Test", url="ws:test", msg={}, ticker="TEST_TEST")
 
 def test_check_quotes_diff_true_case(adapter):
     msg = {'bid':1, 'ask':2}
@@ -76,19 +71,15 @@ def test_valid_message_can_pass_duplicates_validate_false(adapter_validate_messa
     # quotes do differ, in the init function the above is true the values are set to 0
     assert adapter_validate_message_is_false.valid_message_can_pass(msg) == False
 
-def test_folder_creation(adapter):
-    new_path_to_folder = create_base_folder(adapter.exchange_name, adapter.path_to_folder)
-
-    assert os.path.exists(new_path_to_folder)
-    assert os.path.isdir(new_path_to_folder)
-
 
 def test_write_data_to_file(adapter):
-    path_to_folder = os.path.join(adapter.path_to_folder, adapter.exchange_name)
-    for file in os.listdir(path_to_folder):
-        file_path = os.path.join(path_to_folder, file)
-        if os.path.isfile(file_path):
-            os.remove(file_path)
+    path_to_folder = (adapter.PATH_DIR / adapter.exchange_name).resolve()
+    if path_to_folder.exists():
+        for file in os.listdir(path_to_folder):
+            file_path = os.path.join(path_to_folder, file)
+            if os.path.isfile(file_path):
+                os.remove(file_path)
+
     data = [
         {
             "price":1000,
@@ -121,8 +112,6 @@ def test_write_data_to_file(adapter):
 
     if path_to_newly_created_file is None:
         assert False
-    else:
-        assert True
 
     new_df_from_parquet = pd.read_parquet(path_to_newly_created_file)
